% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.plot.R
\name{class.plot}
\alias{class.plot}
\title{A LCGA Model PLotting Function}
\usage{
class.plot(model, data, smooth = "none")
}
\arguments{
\item{model}{enter lcmm model.}

\item{data}{enter data set (complete with multiple ID and predicted class rows).}

\item{smooth}{would you like ggplot to smooth (using loess) time-varying outcome for each class?}
}
\description{
This function helps you plot predicted latent classes generated by an lcmm model.
}
\examples{
class.plot()
}
\keyword{lcmm}
