% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.times.R
\name{def.times}
\alias{def.times}
\title{A Function to Define Time Points}
\usage{
def.times(using_for, start_dose, end_dose)
}
\arguments{
\item{start_dose}{enter starting dosage point.}

\item{end_dose}{enter ending dosage point.}

\item{useing_for}{enter if using function to call time values (e.g., month names), or dosage change points.}
}
\description{
This function is specific to my MSc dissertation, used to more easily define multiple timpoints.
}
\examples{
def.times()
}
\keyword{lcga}
