% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.lcmm.R
\name{evaluate.lcmm}
\alias{evaluate.lcmm}
\title{A LCGA Optimal Model Searching Function}
\usage{
evaluate.lcmm(data, x_name, y_name, g_name, ng_max = 8,
  spline_max = NULL, spline_type = NULL)
}
\arguments{
\item{data}{enter data set with date values converted to day/month/year values.}

\item{x_name}{enter the time-indictating column name (str)}

\item{y_name}{enter the time-varying outcome variable column name (str)}

\item{g_name}{enter the subject grouping variable column name (str; ususally 'ID')}

\item{ng_max}{enter the maximum number of groups you would like to evaluate.}

\item{spline_max}{enter the maximum number of splines you would like the model to evaluate.}

\item{spline_type}{if spline_max != NULL,  enter either the spline type you would like to evaluate (either 'quant' or 'equi')}
}
\description{
This function helps you identify model fit values for different linear, beta, and spline models.
}
\examples{
evaluate.lcmm()
}
\keyword{lcmm}
