% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.since.assign.R
\name{m.since.assign}
\alias{m.since.assign}
\title{A Column Assigning Date to Month Function}
\usage{
m.since.assign(data, date_list, to = "months")
}
\arguments{
\item{data}{enter data set with date columns.}

\item{date_list}{enter date column names (str; using e.g., c("date1", "date2" ...) notation)}

\item{to}{enter the format in which you would like the time difference output (e.g., days, weeks, months, years)}

\item{format}{format the entered date columns should be read as.}
}
\description{
This function will convert date columns to month-since columns (months since first date in time series) and assigns them to a new data frame.
}
\examples{
m.since.assign()
}
\keyword{lcmm}
