% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.since.R
\name{m.since}
\alias{m.since}
\title{A Date/Month Time Difference Function}
\usage{
m.since(late_date, first_date, format = "\%d/\%m/\%Y", to = "months")
}
\arguments{
\item{late_date}{enter data frame column containing the later dates (e.g., second observation in a time-series data set).}

\item{first_date}{enter the earlier date column.}

\item{format}{format the entered date columns should be read as.}

\item{to}{enter the format in which you would like the time difference output (e.g., days, weeks, months, years)}
}
\description{
This function helps you determine the number of months between two columns containing date values.
}
\examples{
m.since()
}
\keyword{lcmm}
