% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.classy.df.R
\name{make.classy.df}
\alias{make.classy.df}
\title{A Data Set Modifying Function}
\usage{
make.classy.df(data, x_name, y_name, p_ids)
}
\arguments{
\item{data}{enter data set containing predicted latent class membership values.}

\item{x_name}{enter the name of the time-indicating column (str)}

\item{y_name}{enter the name of the time-varying outcome column (str)}

\item{p_ids}{enter the name of the column containing the predicted latent class membership values (str)}
}
\description{
This function is used to assist the other functions in the package with iteratively running lcmm models.
}
\examples{
make.classy.df()
}
\keyword{lcmm}
