% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.prelim.df.R
\name{make.prelim.df}
\alias{make.prelim.df}
\title{A Data Set Modifying Function}
\usage{
make.prelim.df(data, x_name, y_name, g_name)
}
\arguments{
\item{data}{enter data set containing predicted latent class membership values.}

\item{x_name}{enter the name of the time-indicating column (str)}

\item{y_name}{enter the name of the time-varying outcome column (str)}

\item{g_name}{enter the name of the column indicating ID values (str)}
}
\description{
This function is used to assist the other functions in the package with iteratively running lcmm models.
}
\examples{
make.prelim.df()
}
\keyword{lcmm}
