% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.time.R
\name{make.time}
\alias{make.time}
\title{A Time-Making Function}
\usage{
make.time(new_data, id_name, time_cutoff)
}
\arguments{
\item{new_data}{enter data for which each ID has the same beginning and end time-point (see 'stop.time').}

\item{id_name}{enter the name (str) of the column in your data containing subject's IDs.}

\item{time_cutoff}{enter the number of days, months, years that define each subject's final observation.}
}
\description{
This function helps you create intermediary time-outcome values for known but unobserved values (e.g., dosages).
}
\examples{
make.time()
}
\keyword{lcmm}
