% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re.plot.R
\name{re.plot}
\alias{re.plot}
\title{A LCGA Model PLotting Function}
\usage{
re.plot(df_stacked, link, ng, smooth = "none")
}
\arguments{
\item{df_stacked}{enter data set with multiple rows for each ID.}

\item{link}{enter link function to use in lcmm model specification.}

\item{ng}{enter number of latent classes you would like to search for.}

\item{smooth}{would you like the plot to smooth (using loess) the time-varying outcome for each latent class?}
}
\description{
This function helps you plot predicted latent classes generated by a newly-specified lcmm model. This function is meant to work in tandem with the other package functions in which columns names for new data sets are specified by the package functions.
}
\examples{
re.plot()
}
\keyword{lcmm}
