% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgmm.R
\name{sgmm}
\alias{sgmm}
\title{SGMM and Inference via Random Scaling}
\usage{
sgmm(
  x,
  y,
  z,
  gamma_0 = 1,
  alpha = 0.501,
  bt_start = NULL,
  inference = "rs",
  n0 = 0,
  Phi_start,
  w_start
)
}
\arguments{
\item{x}{numeric. (n x p) matrix of regressors. Should not include 1 (the intercept)}

\item{y}{numeric}

\item{z}{numeric. (n x q) matrix of instruments. Should not include 1 (the intercept)}

\item{gamma_0}{numeric}

\item{alpha}{numeric}

\item{bt_start}{numeric}

\item{inference}{character specifying the inference method. Default is "rs" (random scaling)}

\item{n0}{numeric}

\item{Phi_start}{(p x q) matrix of starting values for Phi}

\item{w_start}{(q x q) matrix of starting values for W}
}
\value{
#' An object of class \code{"sgdi"}, which is a list containing the following
components:
}
\description{
sgmm computes the SGMM estimator and the confidence interval via random scaling method.
}
\examples{
n = 1e05
p = 5
bt0 = rep(5,p)
x = matrix(rnorm(n*(p-1)), n, (p-1))
y = cbind(1,x) \%*\% bt0 + rnorm(n)
}
