% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to.long.R
\name{to.long}
\alias{to.long}
\title{A Wide to Long Data Set Function}
\usage{
to.long(data, id_column, time_columns, date_columns)
}
\arguments{
\item{data}{enter data set containing month values.}

\item{id_column}{enter the column containing patient IDs (str)}

\item{time_columns}{enter the time-varying outcome variable columns}

\item{date_columns}{enter the time-indicating month-formatted columns (str)}
}
\description{
This function helps you convert a wide time-series data set (with one row for each patient and multiple columns for each different time points), to a long time-series data set (with t rows for each patient where t=number of time poitns).
}
\examples{
to.long()
}
\keyword{lcmm}
